First attempt at fixing HTML output
I suspect that the HTML output is not working in the Docker and
Singularity containers because the graphics device is not being set
correctly. Based on
https://github.com/matplotlib/matplotlib/issues/8929#issuecomment-403475202,
I believe that this issue will be fixed by moving the location where
we set the matplotlib backend to before the import of pyplot.